Add raw_files_to_sync property to SyncFilesConfig
